Handle deletion of models that are linked to products

+++ app/Http/Controllers/ColorController.php
@@ -68,6 +68,10 @@ class ColorController extends Controller
      */
     public function destroy(Color $color)
     {
+        if ($color->productVapes()->exists()) {
+            return redirect()->back()->with('error', 'Cannot delete "' . $color->name . '" because it has products assigned to it.');
+        }
+
         $color->delete();
 
         return redirect()->route('colors.index')
